Add files via upload
Adds a different take on the upgrade path of the bars by making the armoured versions requiring a specific ingredient to be crafted at either the workbench or forge (wood armour at workbench, and iron/steel/stainless armour at the forge) in order to upgrade the bars. Locks those upgrades behind engineering.